Split AuthWorkflow into one workflow per auth command
AuthWorkflow carried ten dependencies for three unrelated commands, so
status and logout had to be handed an OfflineHTTPExecutor — a null
transport that existed only to fill login's seams, and that was then
wired into an OAuth client, a device-authorization factory, and a model
catalog neither command could reach.

AuthLoginWorkflow, AuthStatusWorkflow, and AuthLogoutWorkflow each admit
only what their command uses, so "status opens no socket" is now a
property of the type rather than a runtime promise held up by a null
executor. The store-open failure table the three commands shared moves to
AuthCommandResultMapper.credentialStoreResult so it cannot drift.

Behavior, wording, and exit codes are unchanged.
