+++ Sources/ClawAuth/AuthLoginWorkflow.swift
@@ -22,7 +22,7 @@ public protocol ChatGPTDeviceAuthorizing: Sendable {
 
 // MARK: - Workflow
 
-/// `login`, `status`, and `logout`, with no CLI in them.
+/// `login`, with no CLI in it.
 ///
 /// The ordering inside `login` is the whole point of the type, and it is not a style: the lock comes
 /// before every side effect because a daemon holding the state root must not have a credential
@@ -32,7 +32,7 @@ public protocol ChatGPTDeviceAuthorizing: Sendable {
 /// cannot prepare its secrets should cost the owner nothing, least of all a round trip; and the
 /// profile UUID is minted after the exchange because a login replaces what came before it only once
 /// there is something to replace it with.
-public struct AuthWorkflow: Sendable {
+public struct AuthLoginWorkflow: Sendable {
   private let bootstrap: AuthBootstrap
   private let runtimeSecrets: any AuthRuntimeSecretPreparing
   private let coordinator: AuthMutationCoordinator
@@ -42,7 +42,6 @@ public struct AuthWorkflow: Sendable {
   private let catalog: any ChatGPTModelCatalogFetching
   private let terminal: any AuthTerminal
   private let profileID: @Sendable () -> UUID
-  private let wallDate: @Sendable () -> Date
 
   public init(
     bootstrap: AuthBootstrap,
@@ -53,8 +52,7 @@ public struct AuthWorkflow: Sendable {
     tokenExchange: any ChatGPTOAuthExchanging,
     catalog: any ChatGPTModelCatalogFetching,
     terminal: any AuthTerminal,
-    profileID: @escaping @Sendable () -> UUID,
-    wallDate: @escaping @Sendable () -> Date
+    profileID: @escaping @Sendable () -> UUID
   ) {
     self.bootstrap = bootstrap
     self.runtimeSecrets = runtimeSecrets
@@ -65,10 +63,7 @@ public struct AuthWorkflow: Sendable {
     self.catalog = catalog
     self.terminal = terminal
     self.profileID = profileID
-    self.wallDate = wallDate
   }
-
-  // MARK: - Login
 
   public func login() async -> AuthCommandResult {
     let transcript = AuthTranscript(terminal: terminal)
@@ -83,71 +78,11 @@ public struct AuthWorkflow: Sendable {
       return await transcript.finish(await runLogin(transcript))
     }
   }
-
-  // MARK: - Status
-
-  /// A read, and only a read. No lock, because an owner must be able to ask what the running daemon
-  /// is using without fighting it for the state root; no network and no refresh, because the answer
-  /// to "what is stored" is on the disk, and spending a refresh token to answer it would change the
-  /// thing being reported.
-  public func status() -> AuthCommandResult {
-    let store: any LLMCredentialStore
-    do {
-      store = try makeCredentialStore()
-    } catch let error as LLMCredentialStoreError {
-      return AuthCommandResultMapper.result(for: error)
-    } catch {
-      return AuthCommandResultMapper.unexpected()
-    }
-
-    let stored: StoredOAuthCredential?
-    do {
-      stored = try store.load(providerID: ChatGPTProviderMetadata.providerID)
-    } catch {
-      return AuthCommandResultMapper.result(for: error)
-    }
-
-    var events: [AuthPresentationEvent] = [
-      .output("provider: \(ChatGPTProviderMetadata.providerID.rawValue)")
-    ]
-
-    guard let stored else {
-      events.append(.output("credential: none — logged out. Run `clawd auth login`."))
-      return AuthCommandResult(exit: .success, events: events)
-    }
-
-    events.append(.output("credential: present"))
-    events.append(
-      .output(
-        "expires: \(Self.expiry(stored.expiresAt)) "
-          + "(\(Self.label(for: freshness(of: stored))))"
-      )
-    )
-    if let model = ModelSelection.qualifiedChatGPTModel(in: bootstrap.configuredModel) {
-      events.append(.output("model: \(model)"))
-    }
-    return AuthCommandResult(exit: .success, events: events)
-  }
-
-  // MARK: - Logout
-
-  /// Local deletion, and it says so. Nothing here reaches the vendor, so an access token the owner
-  /// has already been issued stays valid until it expires — an owner who believes otherwise would
-  /// stop looking for a token that is still live.
-  public func logout() -> AuthCommandResult {
-    switch coordinator.acquire() {
-    case .failure(let failure):
-      return AuthCommandResultMapper.result(for: failure)
-    case .success(let lease):
-      defer { lease.release() }
-      return runLogout()
-    }
-  }
 }
 
 // MARK: - The Login Sequence
 
-private extension AuthWorkflow {
+private extension AuthLoginWorkflow {
   /// Streams everything an owner should see through `transcript` as it happens, and returns only the
   /// ending. Splitting it this way is what lets the lock's `defer` in `login` be the one release.
   func runLogin(_ transcript: AuthTranscript) async -> AuthCommandResult {
@@ -203,10 +138,8 @@ private extension AuthWorkflow {
         ),
         providerID: ChatGPTProviderMetadata.providerID
       )
-    } catch let error as LLMCredentialStoreError {
-      return AuthCommandResultMapper.result(for: error)
     } catch {
-      return AuthCommandResultMapper.unexpected()
+      return AuthCommandResultMapper.credentialStoreResult(for: error)
     }
 
     await transcript.emit([
@@ -311,83 +244,9 @@ private extension AuthWorkflow {
   }
 }
 
-// MARK: - The Logout Sequence
-
-private extension AuthWorkflow {
-  func runLogout() -> AuthCommandResult {
-    let store: any LLMCredentialStore
-    do {
-      store = try makeCredentialStore()
-    } catch let error as LLMCredentialStoreError {
-      return AuthCommandResultMapper.result(for: error)
-    } catch {
-      return AuthCommandResultMapper.unexpected()
-    }
-
-    // Asked before it is deleted, so an absent record is answered rather than repaired: `delete`
-    // alone would demand a runtime key this state root may never have had.
-    do {
-      guard try store.load(providerID: ChatGPTProviderMetadata.providerID) != nil else {
-        return AuthCommandResult(
-          exit: .success,
-          events: [
-            .output(
-              "No stored \(ChatGPTProviderMetadata.providerID.rawValue) credential — "
-                + "already logged out."
-            )
-          ]
-        )
-      }
-      try store.delete(providerID: ChatGPTProviderMetadata.providerID)
-    } catch {
-      return AuthCommandResultMapper.result(for: error)
-    }
-
-    return AuthCommandResult(
-      exit: .success,
-      events: [
-        .output("Removed the stored \(ChatGPTProviderMetadata.providerID.rawValue) credential."),
-        .output(
-          """
-          This is a local deletion, not a server-side revocation: an access token already issued \
-          may stay valid until it expires.
-          """
-        ),
-      ]
-    )
-  }
-}
-
 // MARK: - Wording
 
-private extension AuthWorkflow {
-  /// UTC and locale-free: an expiry an owner reads has to mean the same instant as the one the
-  /// vendor issued, whatever the host's region is set to.
-  static func expiry(_ date: Date) -> String {
-    let formatter = ISO8601DateFormatter()
-    formatter.formatOptions = [.withInternetDateTime]
-    formatter.timeZone = TimeZone(secondsFromGMT: 0)
-    return formatter.string(from: date)
-  }
-
-  /// Read through the one classifier the runtime credential source and doctor also read through. A
-  /// second rule here would let an owner be told a token is fresh while the source about to spend it
-  /// decides otherwise.
-  func freshness(of credential: StoredOAuthCredential) -> ChatGPTCredentialFreshness {
-    ChatGPTCredentialFreshness.classify(expiresAt: credential.expiresAt, now: wallDate())
-  }
-
-  static func label(for freshness: ChatGPTCredentialFreshness) -> String {
-    switch freshness {
-    case .fresh:
-      return "fresh"
-    case .expiring:
-      return "expiring"
-    case .expired:
-      return "expired"
-    }
-  }
-
+private extension AuthLoginWorkflow {
   /// The user code, the fixed URL, and the window — and nothing else. The device-auth ID is a bearer
   /// of the pending authorization: anyone holding it can claim the grant the owner is about to
   /// approve, so it belongs on the wire and nowhere an owner or a log can see it.
