+++ docs/ARCHITECTURE.md
@@ -83,7 +83,7 @@ clawd
 | `ClawCore` | lib | Pure domain: value types, config model, **error taxonomy** (§19), **store protocols**, tool/provider/credential/transport/sandbox protocols, the **provider-neutral route, credential, replay-state, accounting, and stream-lifecycle contracts** (§8), the shared **HTTP seam** (`HTTPExecuting`/`HTTPStreaming`/`HTTPResult`, request + response headers — F16), and the 32 768-char `ReplySplitter`. No I/O. | `IncomingMessage`, `SessionKey`, `Principal`, `AppConfig`, `RunState`, `ApprovalState`, `RiskLevel`, `RunBudget`, `HTTPResult`, `HTTPRequest`, `HTTPStreamExchange`, `BoundedAsyncChannel`, `ReplySplitter`, `LLMProviderID`, `LLMProviderDescriptor`, `ResolvedLLMRoute`, `LLMEgressIdentity`, `LLMRequestAuthorization`, `LLMEventStream`, `ProviderExchangeState`, `ProviderCallID`, `LLMCostPolicy`, error taxonomy types; protocols: `LLMProvider`, `LLMCredentialSource`, `LLMCredentialStore`, `ChannelIntake`, `MessageDelivery`, `TelegramTransport` (composite), `HTTPExecuting`, `HTTPStreaming`, `SecretStore`, `ExecutionBackend`, `Tool`, `SearchProviding`, `MessageStore`, `SessionStore`, `RunStore`, `AllowlistStore`, `UpdateCursorStore`, `OutboxStore`, `UsageStore`, `AuditLog`, `MemoryStore`, `ApprovalStore`, `WorkspaceReading` (+ its value types `WorkspaceFile`, `LoadedFile`, `SkillScanResult`) |
 | `ClawData` | lib | GRDB persistence: schema, `DatabaseMigrator`, store implementations. WAL. **Thin `Sendable` wrappers over `any DatabaseWriter`** (not actors guarding the pool); relies on GRDB serialization. | `Database`, concrete `…Store` types conforming to the `ClawCore` protocols |
 | `ClawTelegram` | lib | Thin Bot API client over AsyncHTTPClient; long-poll loop; envelope normalization; **`sendRichMessage` (markdown) + plain `sendMessage` fallback**; `sendChatAction`. (Escaping/splitter live in `ClawCore`.) | `TelegramClient`, `TelegramLongPoller`, `MessageEnvelope`, `InputRichMessage` |
-| `ClawAuth` | lib | ChatGPT protocol metadata + device-code OAuth, model catalog, the refreshable credential actor, and the CLI-neutral auth workflow. Depends only on `ClawCore`, so no auth concept reaches the agent loop. | `ChatGPTProviderMetadata`, `ChatGPTDeviceAuthorization`, `ChatGPTOAuthClient`, `ChatGPTCredentialSource`, `ChatGPTModelCatalog`, `AuthWorkflow`, `AuthBootstrap` |
+| `ClawAuth` | lib | ChatGPT protocol metadata + device-code OAuth, model catalog, the refreshable credential actor, and one CLI-neutral workflow per auth command. Depends only on `ClawCore`, so no auth concept reaches the agent loop. | `ChatGPTProviderMetadata`, `ChatGPTDeviceAuthorization`, `ChatGPTOAuthClient`, `ChatGPTCredentialSource`, `ChatGPTModelCatalog`, `AuthLoginWorkflow`, `AuthStatusWorkflow`, `AuthLogoutWorkflow`, `AuthBootstrap` |
 | `ClawLLM` | lib | **Both wire adapters behind one `LLMProvider`** (§8): OpenAI-compatible Chat Completions + ChatGPT Codex Responses (translation, SSE reconstruction, replay filtering, attempt exposure, retry engine); the OpenAI-shaped message model; the static credential source; usage/cost; retries; **SSE streaming (v1)**; the provider stack factory. Depends on `ClawAuth`. | `OpenAICompatibleProvider`, `ChatGPTResponsesProvider`, `ChatGPTResponsesRequestEncoder`, `ChatGPTResponsesSSEParser`, `ChatGPTProviderStateCodec`, `ProviderAttemptExposure`, `StaticLLMCredentialSource`, `ProviderStackFactory`, `ChatMessage`, `ChatRequest`, `ChatResponse`, `ToolCall`, `Usage`, `CostTable`, `SSEParser` |
 | `ClawSecrets` | lib | Concrete secret backends + the encrypted **provider and MCP credential maps** and their crash-aware publication protocol (§15). | `EncryptedFileSecretStore`, `EnvSecretStore`, `SecretStoreResolver`, `EncryptedLLMCredentialStore`, `EncryptedMCPCredentialStore`, `SecretStatePaths`, `SecureFilePublisher`, `RuntimeSecretPreparer` |
 | `ClawWorkspace` | lib | Identity/memory files (`SOUL/AGENTS/USER/TOOLS/MEMORY.md`, `memory/*.md`, `skills/*`), Yams frontmatter, MCP catalog loading, caps + untrusted-tier injection. | `WorkspaceStore`, `WorkspaceFile`, `MemoryDoc`, `FrontmatterParser`, `MCPConfigLoader` |
@@ -122,7 +122,7 @@ form `ARCHITECTURE.md §N` is used, sparingly.
 | §8 LLM provider | `LLMProvider`, `ChatRequest`/`ChatResponse`, `MessageContent`, `CostResolver`, `UsageResolver` (ClawCore); `ProviderErrorClassifier` (ClawLLM) |
 | §8.1 Route selection | `LLMProviderID`, `LLMProviderDescriptor`, `LLMProviderCapabilities`, `LLMProviderRegistry`, `ResolvedLLMRoute`, `LLMEgressIdentity`, `SessionTraceID` (ClawCore); `ProviderStackFactory`, `ProviderStack` (ClawLLM) |
 | §8.2 Credential seam | `LLMCredentialSource`, `LLMRequestAuthorization`, `LLMCredentialGeneration`, `LLMCredentialRejection`, `LLMCredentialStore`, `StoredOAuthCredential` (ClawCore); `StaticLLMCredentialSource` (ClawLLM); `ChatGPTCredentialSource`, `ChatGPTProviderMetadata`, `ChatGPTTokenMetadata`, `ChatGPTCredentialFreshness` (ClawAuth); `EncryptedLLMCredentialStore` (ClawSecrets) |
-| §8.3 ChatGPT Responses route | `ChatGPTWireValues`, `ChatGPTDeviceAuthorization`, `ChatGPTOAuthClient`, `ChatGPTModelCatalog`, `AuthWorkflow`, `ModelSelection` (ClawAuth); `ChatGPTResponsesProvider`, `ChatGPTResponsesRequestEncoder`, `ChatGPTResponsesSSEParser`, `ChatGPTResponsesAccumulator`, `ChatGPTPromptCacheKey` (ClawLLM) |
+| §8.3 ChatGPT Responses route | `ChatGPTWireValues`, `ChatGPTDeviceAuthorization`, `ChatGPTOAuthClient`, `ChatGPTModelCatalog`, `AuthLoginWorkflow`, `AuthStatusWorkflow`, `AuthLogoutWorkflow`, `ModelSelection` (ClawAuth); `ChatGPTResponsesProvider`, `ChatGPTResponsesRequestEncoder`, `ChatGPTResponsesSSEParser`, `ChatGPTResponsesAccumulator`, `ChatGPTPromptCacheKey` (ClawLLM) |
 | §8.4 Streams & attempt exposure | `LLMEventStream`, `LLMStreamTermination`, `ProviderFailure`, `ProviderFailureAccounting`, `ProviderInferenceCancellation`, `HTTPStreamExchange`, `HTTPTransportFailure`, `BoundedAsyncChannel` (ClawCore); `ProviderAttemptExposure`, `ChatGPTResponsesAttemptEngine` (ClawLLM) |
 | §8.5 Provider replay state | `ProviderExchangeState` (ClawCore); `ChatGPTProviderStateCodec` (ClawLLM) |
 | §8.6 Reliability & cost | `ProviderCallID`, `ProviderCallIDGenerating`, `LLMCostPolicy`, `LLMInputReservationPolicy` (ClawCore); `OpenAICompatibleProvider`, `SSEParser`, `PriceFileLoader` (ClawLLM) |
